requirements.txt
sentence-transformers==2.6.0
langchain
flask
pypdf
python-dotenv
pinecone-client==3.0.0
langchain-pinecone
langchain_community
langchain-google-genai
langchain_experimental
langchain-huggingface
